Move all public programme related func in ProgrammeController
