+++ app/Http/Controllers/ContentModerator/ScheduleController.php
@@ -19,18 +19,6 @@ use Ramsey\Uuid\Type\Time;
 
 class ScheduleController extends Controller
 {
-    public function index(): View
-    {
-        if (!EventInstance::current()->is_final_programme_released)
-            abort(404);
-
-        $lectureTimeslots = Timeslot::where('duration', 30)->get();
-        $workshopTimeslots = Timeslot::where('duration', 90)->get();
-
-        return view('presentations.index',
-            compact('lectureTimeslots', 'workshopTimeslots'));
-    }
-
     /**
      * Display the overview for scheduling
      *
